add digital diplay with chart, update pi_controller with integrate analog, scanning algorithm read from data global store directly,

+++ CMakeProject2/include/scanning/scanning_algorithm.cpp
@@ -207,7 +207,7 @@ bool ScanningAlgorithm::ExecuteScanSequence() {
 
       // Calculate progress
       size_t currentStep = axisIndex + stepIndex * m_parameters.axesToScan.size();
-      double totalSteps = m_parameters.axesToScan.size() * m_parameters.stepSizes.size();
+      double totalSteps = static_cast<double>(m_parameters.axesToScan.size() * m_parameters.stepSizes.size());
       double progress = static_cast<double>(currentStep) / totalSteps;
 
       std::stringstream progressMsg;
